split up control.ts to allow for overriding top level functions
