+++ libs/base/control.ts
@@ -253,197 +253,3 @@ namespace control {
         }
     }
 }
-
-/**
- * Busy wait for a condition to be true
- * @param condition condition to test for
- * @param timeOut if positive, maximum duration to wait for in milliseconds
- */
-//% blockId="pxt_pause_until"
-function pauseUntil(condition: () => boolean, timeOut?: number): void {
-    if (!condition || condition()) return; // optimistic path
-    if (!timeOut) timeOut = 0;
-    control.__queuePollEvent(timeOut, condition, undefined);
-}
-
-/**
- * Repeats the code forever in the background. On each iteration, allows other codes to run.
- * @param body code to execute
- */
-//% help=loops/forever weight=100 afterOnStart=true blockNamespace="loops"
-//% blockId=forever block="forever" blockAllowMultiple=1
-function forever(a: () => void): void {
-    loops.forever(a);
-}
-
-/**
- * Pause for the specified time in milliseconds
- * @param ms how long to pause for, eg: 100, 200, 500, 1000, 2000
- */
-//% help=loops/pause weight=99
-//% async block="pause %pause=timePicker|ms"
-//% blockId=device_pause blockNamespace="loops"
-function pause(ms: number): void {
-    loops.pause(ms);
-}
-
-/**
- * Convert any value to text
- * @param value value to be converted to text
- */
-//% help=text/convert-to-text weight=1
-//% block="convert $value=math_number to text"
-//% blockId=variable_to_text blockNamespace="text"
-function convertToText(value: any): string {
-    return "" + value;
-}
-
-// micro:bit compatibility
-// these functions allow some level of reuse
-// between micro:bit and other maker-style editors
-namespace basic {
-    export function pause(millis: number) {
-        loops.pause(millis);
-    }
-}
-
-/**
- * Calls a function with a fixed time delay between each call to that function.
- * @param func 
- * @param delay 
- */
-//%
-function setInterval(func: () => void, delay: number): number {
-    delay = Math.max(10, delay | 0);
-    return control.setInterval(func, delay, control.IntervalMode.Interval);
-}
-
-/**
- * Cancels repeated action which was set up using setInterval().
- * @param intervalId 
- */
-//%
-function clearInterval(intervalId: number) {
-    control.clearInterval(intervalId, control.IntervalMode.Interval);
-}
-
-/**
- * Calls a function after specified delay.
- * @param func 
- * @param delay 
- */
-//%
-function setTimeout(func: () => void, delay: number): number {
-    return control.setInterval(func, delay, control.IntervalMode.Timeout);
-}
-
-/**
- * Clears the delay set by setTimeout().
- * @param intervalId 
- */
-//%
-function clearTimeout(intervalId: number) {
-    control.clearInterval(intervalId, control.IntervalMode.Timeout);
-}
-
-/**
- * Calls a function as soon as possible.
- * @param func 
- */
-//%
-function setImmediate(func: () => void): number {
-    return control.setInterval(func, 0, control.IntervalMode.Immediate);
-}
-
-/**
- * Cancels the immediate actions.
- * @param intervalId 
- */
-//%
-function clearImmediate(intervalId: number) {
-    control.clearInterval(intervalId, control.IntervalMode.Immediate);
-}
-
-class UTF8Decoder {
-    private buf: Buffer;
-
-    constructor() {
-        this.buf = undefined;
-    }
-
-    add(buf: Buffer) {
-        if (!buf || !buf.length) return;
-
-        if (!this.buf)
-            this.buf = buf;
-        else {
-            const b = control.createBuffer(this.buf.length + buf.length);
-            b.write(0, this.buf);
-            b.write(this.buf.length, buf);
-            this.buf = b;
-        }
-    }
-
-    decodeUntil(delimiter: number): string {
-        if (!this.buf) return undefined;
-        delimiter = delimiter | 0;
-        let i = 0;
-        for (; i < this.buf.length; ++i) {
-            const c = this.buf[i];
-            // skip multi-chars
-            if ((c & 0xe0) == 0xc0)
-                i += 1;
-            else if ((c & 0xf0) == 0xe0)
-                i += 2;
-            else if (c == delimiter) {
-                // found it
-                break;
-            }
-        }
-
-        if (i >= this.buf.length)
-            return undefined;
-        else {
-            const s = this.buf.slice(0, i).toString();
-            if (i + 1 == this.buf.length)
-                this.buf = undefined;
-            else
-                this.buf = this.buf.slice(i + 1);
-            return s;
-        }
-    }
-
-    decode(): string {
-        if (!this.buf) return "";
-
-        // scan the end of the buffer for partial characters
-        let length = 0;
-        for (let i = this.buf.length - 1; i >= 0; i--) {
-            const c = this.buf[i];
-            if ((c & 0x80) == 0) {
-                length = i + 1;
-                break;
-            }
-            else if ((c & 0xe0) == 0xc0) {
-                length = i + 2;
-                break;
-            }
-            else if ((c & 0xf0) == 0xe0) {
-                length = i + 3;
-                break;
-            }
-        }
-        // is last beyond the end?
-        if (length == this.buf.length) {
-            const s = this.buf.toString();
-            this.buf = undefined;
-            return s;
-        } else if (length == 0) { // data yet
-            return "";
-        } else {
-            const s = this.buf.slice(0, length).toString();
-            this.buf = this.buf.slice(length);
-            return s;
-        }
-    }
-}
